requirement.txt
python 3.7.11
pytorch 1.7.1
numpy 1.20.3
